requirements.txt
pyotp
mysql-connector-python
hvac